add dialog as separate comp
